properly fix 32/64 bit confusion (make idxint int instead of long to match SuiteSparse usage), fix all remaining warning messages

+++ external/amd/src/amd_order.c
@@ -183,7 +183,7 @@ GLOBAL Int AMD_order
     /* order the matrix */
     /* --------------------------------------------------------------------- */
 
-    AMD_1 (n, Cp, Ci, P, Pinv, Len, slen, S, Control, Info) ;
+    AMD_1 (n, Cp, Ci, P, Pinv, Len, (Int)slen, S, Control, Info) ;
 
     /* --------------------------------------------------------------------- */
     /* free the workspace */
